improved cache helper to protect from returning deleted/disposed entities. Refactored code of cache, removed unused leftovers, unit tested.

+++ packages/dev/occt-worker/lib/occ-worker/cache-helper.ts
@@ -10,17 +10,35 @@ export class CacheHelper {
     constructor(private readonly occ: OpenCascadeInstance) { }
 
     cleanAllCache(): void {
-        const usedHashKeys = Object.keys(this.usedHashes);
+        // Clean all entries in argCache, not just usedHashes
+        const allCacheKeys = Object.keys(this.argCache);
 
-        usedHashKeys.forEach(hash => {
+        allCacheKeys.forEach(hash => {
             if (this.argCache[hash]) {
                 try {
-                    const shape = this.argCache[hash];
-                    this.occ.BRepTools.Clean(shape, true);
-                    this.occ.BRepTools.CleanGeometry(shape);
-                    shape.delete();
+                    const cachedItem = this.argCache[hash];
+                    // Only attempt to clean and delete OCCT shapes
+                    if (this.isOCCTObject(cachedItem)) {
+                        // Handle arrays of OCCT objects
+                        if (Array.isArray(cachedItem)) {
+                            cachedItem.forEach(shape => {
+                                try {
+                                    this.occ.BRepTools.Clean(shape, true);
+                                    this.occ.BRepTools.CleanGeometry(shape);
+                                    shape.delete();
+                                } catch (error) {
+                                    // Ignore errors for already deleted shapes
+                                }
+                            });
+                        } else {
+                            this.occ.BRepTools.Clean(cachedItem, true);
+                            this.occ.BRepTools.CleanGeometry(cachedItem);
+                            cachedItem.delete();
+                        }
+                    }
                 }
-                catch {
+                catch (error) {
+                    // Ignore errors when cleaning shapes that may already be deleted
                 }
             }
         });
@@ -33,12 +51,29 @@ export class CacheHelper {
     cleanCacheForHash(hash: string): void {
         if (this.argCache[hash]) {
             try {
-                const shape = this.argCache[hash];
-                this.occ.BRepTools.Clean(shape, true);
-                this.occ.BRepTools.CleanGeometry(shape);
-                shape.delete();
+                const cachedItem = this.argCache[hash];
+                // Only attempt to clean and delete OCCT shapes
+                if (this.isOCCTObject(cachedItem)) {
+                    // Handle arrays of OCCT objects
+                    if (Array.isArray(cachedItem)) {
+                        cachedItem.forEach(shape => {
+                            try {
+                                this.occ.BRepTools.Clean(shape, true);
+                                this.occ.BRepTools.CleanGeometry(shape);
+                                shape.delete();
+                            } catch (error) {
+                                // Ignore errors for already deleted shapes
+                            }
+                        });
+                    } else {
+                        this.occ.BRepTools.Clean(cachedItem, true);
+                        this.occ.BRepTools.CleanGeometry(cachedItem);
+                        cachedItem.delete();
+                    }
+                }
             }
-            catch {
+            catch (error) {
+                // Ignore errors when cleaning shapes that may already be deleted
             }
         }
         delete this.argCache[hash];
@@ -47,37 +82,59 @@ export class CacheHelper {
     }
 
     cleanUpCache(): void {
-        // TODO seems to have problems, not exactly sure what is the real cause. For now users will build up cache during the whole session.
-        // and will be able to manually clean it up if it gets out of hand.
-
-
-        // const usedHashKeys = Object.keys(this.usedHashes);
-        // const hashesFromPreviousRunKeys = Object.keys(this.hashesFromPreviousRun);
-        // const newArgsCache = {};
-        // const newUsedHashKeys = {};
-
-        // let hashesToDelete = [];
-        // if (usedHashKeys.length > 0 && hashesFromPreviousRunKeys.length > 0) {
-        //     hashesToDelete = usedHashKeys.filter(hash => !hashesFromPreviousRunKeys.includes(hash));
-        // }
-        // hashesFromPreviousRunKeys.forEach(hash => {
-        //     newArgsCache[hash] = this.argCache[hash];
-        //     newUsedHashKeys[hash] = this.usedHashes[hash];
-        // });
-
-        // if (hashesToDelete.length > 0) {
-        //     hashesToDelete.forEach(hash => {
-        //         if (this.argCache[hash]) {
-        //             this.argCache[hash].delete();
-        //         }
-        //     });
-        // }
-        // this.argCache = newArgsCache;
-        // this.usedHashes = newUsedHashKeys;
-        // this.hashesFromPreviousRun = {};
+        // Clean up cache entries that were used in previous run but not in current run
+        // This helps manage memory by removing unused cached shapes
+        
+        const usedHashKeys = Object.keys(this.usedHashes);
+        const hashesFromPreviousRunKeys = Object.keys(this.hashesFromPreviousRun);
+        
+        // Find hashes that exist in previous run but not in current run
+        // These are the ones we should clean up
+        let hashesToDelete: string[] = [];
+        if (hashesFromPreviousRunKeys.length > 0) {
+            hashesToDelete = hashesFromPreviousRunKeys.filter(hash => !usedHashKeys.includes(hash));
+        }
+        
+        // Delete unused shapes and clean them from cache
+        if (hashesToDelete.length > 0) {
+            hashesToDelete.forEach(hash => {
+                if (this.argCache[hash]) {
+                    try {
+                        const shape = this.argCache[hash];
+                        // Only try to clean and delete if it's an OCCT object
+                        if (this.isOCCTObject(shape)) {
+                            // Handle arrays of OCCT objects
+                            if (Array.isArray(shape)) {
+                                shape.forEach(s => {
+                                    try {
+                                        this.occ.BRepTools.Clean(s, true);
+                                        this.occ.BRepTools.CleanGeometry(s);
+                                        s.delete();
+                                    } catch {
+                                        // Ignore errors for already deleted shapes
+                                    }
+                                });
+                            } else {
+                                this.occ.BRepTools.Clean(shape, true);
+                                this.occ.BRepTools.CleanGeometry(shape);
+                                shape.delete();
+                            }
+                        }
+                    } catch {
+                        // Ignore errors for already deleted or invalid shapes
+                    }
+                    delete this.argCache[hash];
+                }
+                delete this.usedHashes[hash];
+            });
+        }
+        
+        // Update hashesFromPreviousRun to be current usedHashes for next cleanup cycle
+        this.hashesFromPreviousRun = { ...this.usedHashes };
     }
 
-    isOCCTObject(obj): boolean {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    isOCCTObject(obj: any): boolean {
         return obj !== undefined && obj !== null && (!Array.isArray(obj) && obj.$$ !== undefined) || (Array.isArray(obj) && obj.length > 0 && obj[0].$$ !== undefined);
     }
 
@@ -86,15 +143,14 @@ export class CacheHelper {
      * call the `cacheMiss()` callback otherwise. The result will be
      * added to the cache if `GUIState["Cache?"]` is true.
      */
-    cacheOp(args, cacheMiss): any {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    cacheOp(args: any, cacheMiss: () => any): any {
         let toReturn = null;
         const curHash = this.computeHash(args);
         this.usedHashes[curHash] = curHash;
         this.hashesFromPreviousRun[curHash] = curHash;
         const check = this.checkCache(curHash);
         if (check) {
-            // TODO I need to check if and why cloning is required.
-            // toReturn = new this.occ.TopoDS_Shape(check);
             if (this.isOCCTObject(check)) {
                 toReturn = check;
                 toReturn.hash = check.hash;
@@ -109,32 +165,32 @@ export class CacheHelper {
                     r.hash = itemHash;
                     this.addToCache(itemHash, r);
                 });
-            } else {
-                if (this.isOCCTObject(toReturn)) {
-                    toReturn.hash = curHash;
-                    this.addToCache(curHash, toReturn);
-                } else if (toReturn && toReturn.compound && toReturn.data && toReturn.shapes && toReturn.shapes.length > 0) {
-                    const objDef: Models.OCCT.ObjectDefinition<any, any> = toReturn;
-                    const compoundHash = this.computeHash({ ...args, index: "compound" });
-                    objDef.compound.hash = compoundHash;
-                    this.addToCache(compoundHash, objDef.compound);
-                    objDef.shapes.forEach((s, index) => {
-                        const itemHash = this.computeHash({ ...args, index });
-                        s.shape.hash = itemHash;
-                        this.addToCache(itemHash, s.shape);
-                    });
-                    this.addToCache(curHash, { value: objDef });
-                }
-                else {
-                    this.addToCache(curHash, { value: toReturn });
-                }
-            }
-
-        }
+                } else {
+                    if (this.isOCCTObject(toReturn)) {
+                        toReturn.hash = curHash;
+                        this.addToCache(curHash, toReturn);
+                    } else if (toReturn && toReturn.compound && toReturn.data && toReturn.shapes && toReturn.shapes.length > 0) {
+                        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                        const objDef: Models.OCCT.ObjectDefinition<any, any> = toReturn;
+                        const compoundHash = this.computeHash({ ...args, index: "compound" });
+                        objDef.compound.hash = compoundHash;
+                        this.addToCache(compoundHash, objDef.compound);
+                        objDef.shapes.forEach((s, index) => {
+                            const itemHash = this.computeHash({ ...args, index });
+                            s.shape.hash = itemHash;
+                            this.addToCache(itemHash, s.shape);
+                        });
+                        this.addToCache(curHash, { value: objDef });
+                    }
+                    else {
+                        this.addToCache(curHash, { value: toReturn });
+                    }
+                }        }
         return toReturn;
     }
     /** Returns the cached object if it exists and is valid, or null otherwise. */
-    checkCache(hash): any {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    checkCache(hash: string | number): any {
         const cachedShape = this.argCache[hash];
         if (!cachedShape) {
             return null;
@@ -180,10 +236,13 @@ export class CacheHelper {
         return cachedShape;
     }
     /** Adds this `shape` to the cache, indexable by `hash`. */
-    addToCache(hash, shape): any {
-        // TODO I need to check if and why casting is required. Having real objects in the cache can be used to free up memory?
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    addToCache(hash: string | number, shape: any): string | number {
         const cacheShape = shape;
-        cacheShape.hash = hash; // This is the cached version of the object
+        // Only set hash property on objects, not primitives
+        if (cacheShape !== null && typeof cacheShape === "object") {
+            cacheShape.hash = hash; // This is the cached version of the object
+        }
         this.argCache[hash] = cacheShape;
         return hash;
     }
@@ -191,10 +250,11 @@ export class CacheHelper {
     /** This function computes a 32-bit integer hash given a set of `arguments`.
      * If `raw` is true, the raw set of sanitized arguments will be returned instead.
      */
-    computeHash(args, raw?: any): any {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    computeHash(args: any, raw?: boolean): number | string {
         let argsString = JSON.stringify(args);
-        argsString = argsString.replace(/(\"ptr\"\:(-?[0-9]*?)\,)/g, "");
-        argsString = argsString.replace(/(\"ptr\"\:(-?[0-9]*))/g, "");
+        argsString = argsString.replace(/("ptr":(-?[0-9]*?),)/g, "");
+        argsString = argsString.replace(/("ptr":(-?[0-9]*))/g, "");
         if (argsString.includes("ptr")) { console.error("YOU DONE MESSED UP YOUR REGEX."); }
         const hashString = Math.random.toString() + argsString;
         if (raw) { return hashString; }
@@ -202,7 +262,7 @@ export class CacheHelper {
     }
 
     /** This function converts a string to a 32bit integer. */
-    stringToHash(str: string): any {
+    stringToHash(str: string): number {
         let hash = 0;
         if (str.length === 0) { return hash; }
         for (let i = 0; i < str.length; i++) {
@@ -216,31 +276,12 @@ export class CacheHelper {
     }
 
     /** This function returns a version of the `inputArray` without the `objectToRemove`. */
-    remove(inputArray, objectToRemove): any {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    remove(inputArray: any[], objectToRemove: any): any[] {
         return inputArray.filter((el) => {
             return el.hash !== objectToRemove.hash ||
                 el.ptr !== objectToRemove.ptr;
         });
-    }
-
-    dupShape(shape): any {
-        if (shape.ShapeType() === this.occ.TopAbs_ShapeEnum.TopAbs_WIRE) {
-            return this.occ.TopoDS.Wire_1(shape);
-        } else if (shape.ShapeType() === this.occ.TopAbs_ShapeEnum.TopAbs_SHELL) {
-            return this.occ.TopoDS.Shell_1(shape);
-        } else if (shape.ShapeType() === this.occ.TopAbs_ShapeEnum.TopAbs_EDGE) {
-            return this.occ.TopoDS.Edge_1(shape);
-        } else if (shape.ShapeType() === this.occ.TopAbs_ShapeEnum.TopAbs_SOLID) {
-            return this.occ.TopoDS.Solid_1(shape);
-        } else if (shape.ShapeType() === this.occ.TopAbs_ShapeEnum.TopAbs_FACE) {
-            return this.occ.TopoDS.Face_1(shape);
-        } else if (shape.ShapeType() === this.occ.TopAbs_ShapeEnum.TopAbs_COMPOUND) {
-            return this.occ.TopoDS.Compound_1(shape);
-        } else if (shape.ShapeType() === this.occ.TopAbs_ShapeEnum.TopAbs_COMPSOLID) {
-            return this.occ.TopoDS.CompSolid_1(shape);
-        } else if (shape.ShapeType() === this.occ.TopAbs_ShapeEnum.TopAbs_VERTEX) {
-            return this.occ.TopoDS.Vertex_1(shape);
-        }
     }
 
 }
